requirements.txt
fastapi
uvicorn[standard]
roboflow
pillow
python-dotenv
pandas
numpy
supabase
slowapi
python-multipart